Topology: fix AllLoopedPairs, AllIsolatedBranches, AllIsolatedLoads
Especially in AllLoopedPairs, the array is resized in the loop and there are comparisons
with the old values. To avoid multiple conversions in the loop, I replaced the pointers
with Pascal strings and dynamic arrays. At the end of the procedure, everything is
copied to the ouput pointers as usual.